# BackEnd/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "马术障碍赛路线设计器 API",
        "version": "1.0.0",
        "description": "路线设计、用户、会员、支付、AI 生成、反馈和协作接口"
    },
    "paths": {
...
        "/user/admin/analytics/": {
            "get": {
                "operationId": "admin_analytics",
                "description": "运营数据看板核心指标。",
                "tags": [
                    "user"
                ],
                "security": [
                    {
                        "CookieJWTAuth": []
                    },
                    {
                        "jwtAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {}
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "CookieJWTAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token",
                "description": "登录后由后端写入的 httpOnly JWT Cookie"
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}